fix(sdk): don't expose private methods in Aztec
